Split host storage reset boundary coverage
